+++ src/main/java/me/drex/worldmanager/WorldManager.java
@@ -1,6 +1,9 @@
 package me.drex.worldmanager;
 
+import eu.pb4.playerdata.api.PlayerDataApi;
+import eu.pb4.playerdata.api.storage.NbtCodecDataStorage;
 import me.drex.worldmanager.command.WorldManagerCommand;
+import me.drex.worldmanager.data.PlayerData;
 import me.drex.worldmanager.save.ChunkGenerators;
 import me.drex.worldmanager.save.WorldManagerSavedData;
 import net.fabricmc.api.ModInitializer;
@@ -14,9 +17,11 @@ public class WorldManager implements ModInitializer {
 
     public static final String MOD_ID = "worldmanager";
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
+    public static final NbtCodecDataStorage<PlayerData> STORAGE = new NbtCodecDataStorage<>("worldmanager", PlayerData.CODEC);
 
     @Override
     public void onInitialize() {
+        PlayerDataApi.register(STORAGE);
         ChunkGenerators.init();
 
         ServerLifecycleEvents.SERVER_STARTED.register(server -> {