+++ src/main/java/me/drex/worldmanager/command/TeleportCommand.java
@@ -2,10 +2,15 @@ package me.drex.worldmanager.command;
 
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import eu.pb4.playerdata.api.PlayerDataApi;
+import me.drex.worldmanager.WorldManager;
+import me.drex.worldmanager.data.PlayerData;
+import me.drex.worldmanager.save.Location;
 import me.drex.worldmanager.save.WorldConfig;
 import me.drex.worldmanager.save.WorldManagerSavedData;
 import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.commands.Commands;
 import net.minecraft.commands.arguments.EntityArgument;
 import net.minecraft.commands.arguments.ResourceLocationArgument;
 import net.minecraft.core.registries.Registries;
@@ -26,22 +31,22 @@ import static me.drex.worldmanager.command.WorldManagerCommand.WORLD_SUGGESTIONS
 import static net.minecraft.commands.Commands.argument;
 import static net.minecraft.commands.Commands.literal;
 
-public class SpawnCommand {
+public class TeleportCommand {
     public static LiteralArgumentBuilder<CommandSourceStack> build() {
-        return literal("spawn")
-            .requires(Permissions.require("worldmanager.command.worldmanager.spawn", 2))
+        return literal("tp")
+            .requires(Permissions.require("worldmanager.command.worldmanager.teleport", 2))
             .then(
                 argument("id", ResourceLocationArgument.id())
                     .suggests(WORLD_SUGGESTIONS)
-                    .executes(context -> spawn(context.getSource(), ResourceLocationArgument.getId(context, "id"), List.of(context.getSource().getPlayerOrException())))
+                    .executes(context -> teleport(context.getSource(), ResourceLocationArgument.getId(context, "id"), List.of(context.getSource().getPlayerOrException())))
                     .then(
                         argument("targets", EntityArgument.players())
-                            .executes(context -> spawn(context.getSource(), ResourceLocationArgument.getId(context, "id"), EntityArgument.getPlayers(context, "targets")))
+                            .executes(context -> teleport(context.getSource(), ResourceLocationArgument.getId(context, "id"), EntityArgument.getPlayers(context, "targets")))
                     )
             );
     }
 
-    public static int spawn(CommandSourceStack source, ResourceLocation id, Collection<ServerPlayer> targets) throws CommandSyntaxException {
+    public static int teleport(CommandSourceStack source, ResourceLocation id, Collection<ServerPlayer> targets) throws CommandSyntaxException {
         ResourceKey<Level> resourceKey = ResourceKey.create(Registries.DIMENSION, id);
         MinecraftServer server = source.getServer();
         WorldManagerSavedData savedData = WorldManagerSavedData.getSavedData(server);
@@ -51,12 +56,22 @@ public class SpawnCommand {
             throw UNKNOWN_WORLD.create();
         }
         for (ServerPlayer player : targets) {
-            TeleportTransition teleportTransition = config.data.spawnLocation
-                .map(location -> location.toTeleportTransition(serverLevel))
-                .orElseGet(() -> TeleportTransition.missingRespawnBlock(serverLevel, player, TeleportTransition.DO_NOTHING));
+            PlayerData playerData = PlayerDataApi.getCustomDataFor(player, WorldManager.STORAGE);
+            TeleportTransition teleportTransition = null;
+            if (playerData != null) {
+                Location lastLocation = playerData.locations().get(resourceKey);
+                if (lastLocation != null) {
+                    teleportTransition = lastLocation.toTeleportTransition(serverLevel);
+                }
+            }
+            if (teleportTransition == null) {
+                teleportTransition = config.data.spawnLocation
+                    .map(location -> location.toTeleportTransition(serverLevel))
+                    .orElseGet(() -> TeleportTransition.missingRespawnBlock(serverLevel, player, TeleportTransition.DO_NOTHING));
+            }
             player.teleport(teleportTransition);
         }
-        source.sendSuccess(() -> builder("worldmanager.command.spawn").addPlaceholder("id", id.toString()).build(), false);
+        source.sendSuccess(() -> builder("worldmanager.command.teleport").addPlaceholder("id", id.toString()).build(), false);
         return 1;
     }
 }